+++ src/Contracts/Factories/FactoryInterface.php
@@ -19,7 +19,6 @@
 namespace CloudCreativity\JsonApi\Contracts\Factories;
 
 use CloudCreativity\JsonApi\Contracts\Encoder\SerializerInterface;
-use CloudCreativity\JsonApi\Contracts\Http\ApiInterface;
 use CloudCreativity\JsonApi\Contracts\Http\Client\ClientInterface;
 use CloudCreativity\JsonApi\Contracts\Http\Requests\RequestInterface;
 use CloudCreativity\JsonApi\Contracts\Http\Requests\RequestInterpreterInterface;
@@ -33,15 +32,16 @@ use CloudCreativity\JsonApi\Contracts\Utils\ReplacerInterface;
 use CloudCreativity\JsonApi\Contracts\Validators\QueryValidatorInterface;
 use CloudCreativity\JsonApi\Contracts\Validators\ValidatorFactoryInterface;
 use Neomerx\JsonApi\Contracts\Codec\CodecMatcherInterface;
+use Neomerx\JsonApi\Contracts\Document\ErrorInterface;
 use Neomerx\JsonApi\Contracts\Document\LinkInterface;
 use Neomerx\JsonApi\Contracts\Factories\FactoryInterface as BaseFactoryInterface;
-use Neomerx\JsonApi\Contracts\Http\Headers\SupportedExtensionsInterface;
 use Neomerx\JsonApi\Contracts\Http\Query\QueryCheckerInterface;
 use Neomerx\JsonApi\Contracts\Schema\ContainerInterface as SchemaContainerInterface;
 use Neomerx\JsonApi\Encoder\EncoderOptions;
+use Neomerx\JsonApi\Exceptions\ErrorCollection;
 use Psr\Http\Message\MessageInterface;
-use Psr\Http\Message\ServerRequestInterface;
 use Psr\Http\Message\ResponseInterface as PsrResponse;
+use Psr\Http\Message\ServerRequestInterface;
 
 /**
  * Interface FactoryInterface
@@ -60,26 +60,6 @@ interface FactoryInterface extends BaseFactoryInterface
      * @return SerializerInterface
      */
     public function createSerializer(SchemaContainerInterface $container, EncoderOptions $encoderOptions = null);
-
-    /**
-     * @param $namespace
-     * @param CodecMatcherInterface $codecMatcher
-     * @param SchemaContainerInterface $schemaContainer
-     * @param StoreInterface $store
-     * @param ErrorRepositoryInterface $errorRepository
-     * @param SupportedExtensionsInterface|null $supportedExtensions
-     * @param string|null $urlPrefix
-     * @return ApiInterface
-     */
-    public function createApi(
-        $namespace,
-        CodecMatcherInterface $codecMatcher,
-        SchemaContainerInterface $schemaContainer,
-        StoreInterface $store,
-        ErrorRepositoryInterface $errorRepository,
-        SupportedExtensionsInterface $supportedExtensions = null,
-        $urlPrefix = null
-    );
 
     /**
      * Create a JSON API request object from a PSR server request.
@@ -105,6 +85,15 @@ interface FactoryInterface extends BaseFactoryInterface
      * @return ResponseInterface
      */
     public function createResponse(PsrResponse $response);
+
+    /**
+     * Create an error response object.
+     *
+     * @param ErrorInterface|ErrorInterface[]|ErrorCollection $errors
+     * @param int|null $defaultHttpCode
+     * @param array $headers
+     */
+    public function createErrorResponse($errors, $defaultHttpCode, array $headers = []);
 
     /**
      * Create a JSON API document from a HTTP message.