+++ src/Factories/Factory.php
@@ -20,17 +20,16 @@ namespace CloudCreativity\JsonApi\Factories;
 
 use CloudCreativity\JsonApi\Contracts\Encoder\SerializerInterface;
 use CloudCreativity\JsonApi\Contracts\Factories\FactoryInterface;
-use CloudCreativity\JsonApi\Contracts\Http\ApiInterface;
 use CloudCreativity\JsonApi\Contracts\Http\Requests\RequestInterpreterInterface;
 use CloudCreativity\JsonApi\Contracts\Repositories\ErrorRepositoryInterface;
 use CloudCreativity\JsonApi\Contracts\Store\ContainerInterface as AdapterContainerInterface;
 use CloudCreativity\JsonApi\Contracts\Store\StoreInterface;
 use CloudCreativity\JsonApi\Contracts\Validators\QueryValidatorInterface;
 use CloudCreativity\JsonApi\Encoder\Encoder;
-use CloudCreativity\JsonApi\Http\Api;
 use CloudCreativity\JsonApi\Http\Client\GuzzleClient;
 use CloudCreativity\JsonApi\Http\Query\ValidationQueryChecker;
 use CloudCreativity\JsonApi\Http\Requests\RequestFactory;
+use CloudCreativity\JsonApi\Http\Responses\ErrorResponse;
 use CloudCreativity\JsonApi\Http\Responses\Response;
 use CloudCreativity\JsonApi\Object\Document;
 use CloudCreativity\JsonApi\Pagination\Page;
@@ -41,11 +40,11 @@ use CloudCreativity\JsonApi\Store\Store;
 use CloudCreativity\JsonApi\Utils\Replacer;
 use CloudCreativity\JsonApi\Validators\ValidatorErrorFactory;
 use CloudCreativity\JsonApi\Validators\ValidatorFactory;
-use Neomerx\JsonApi\Contracts\Codec\CodecMatcherInterface;
+use Neomerx\JsonApi\Contracts\Document\ErrorInterface;
 use Neomerx\JsonApi\Contracts\Document\LinkInterface;
-use Neomerx\JsonApi\Contracts\Http\Headers\SupportedExtensionsInterface;
 use Neomerx\JsonApi\Contracts\Schema\ContainerInterface as SchemaContainerInterface;
 use Neomerx\JsonApi\Encoder\EncoderOptions;
+use Neomerx\JsonApi\Exceptions\ErrorCollection;
 use Neomerx\JsonApi\Factories\Factory as BaseFactory;
 use Psr\Http\Message\MessageInterface;
 use Psr\Http\Message\ResponseInterface as PsrResponse;
@@ -83,29 +82,6 @@ class Factory extends BaseFactory implements FactoryInterface
     /**
      * @inheritDoc
      */
-    public function createApi(
-        $namespace,
-        CodecMatcherInterface $codecMatcher,
-        SchemaContainerInterface $schemaContainer,
-        StoreInterface $store,
-        ErrorRepositoryInterface $errorRepository,
-        SupportedExtensionsInterface $supportedExtensions = null,
-        $urlPrefix = null
-    ) {
-        return new Api(
-            $namespace,
-            $codecMatcher,
-            $schemaContainer,
-            $store,
-            $errorRepository,
-            $supportedExtensions,
-            $urlPrefix
-        );
-    }
-
-    /**
-     * @inheritDoc
-     */
     public function createRequest(
         ServerRequestInterface $httpRequest,
         RequestInterpreterInterface $intepreter,
@@ -123,6 +99,14 @@ class Factory extends BaseFactory implements FactoryInterface
     public function createResponse(PsrResponse $response)
     {
         return new Response($response, $this->createDocumentObject($response));
+    }
+
+    /**
+     * @inheritDoc
+     */
+    public function createErrorResponse($errors, $defaultHttpCode, array $headers = [])
+    {
+        return new ErrorResponse($errors, $defaultHttpCode, $headers);
     }
 
     /**